feat (fe): removed course dashboard filter and moved it to course dashboard header instead
