+++ frontend/src/features/courses/dashboard/course-dashboard-header.tsx
@@ -1,6 +1,5 @@
-import type { FilterType } from '@/components/multi-filter.tsx'
+import { MultiFilter, type FilterType } from '@/components/multi-filter.tsx'
 import SearchComponent from '@/components/search-component.tsx'
-import CourseDashboardFilters from '@/features/courses/dashboard/course-dashboard-filters.tsx'
 import {
   Box,
   Button,
@@ -17,7 +16,7 @@ import {
   IconLayoutGridFilled,
   IconList,
 } from '@tabler/icons-react'
-import { type ReactNode } from 'react'
+import { useMemo, type ReactNode } from 'react'
 
 type DashboardHeaderProps = {
   coursesData: any
@@ -31,7 +30,7 @@ type DashboardHeaderProps = {
   onViewChange: (view: 'grid' | 'list') => void
 }
 
-const CourseDashboardHeader = ({
+function CourseDashboardHeader({
   coursesData,
   filters,
   activeFilters,
@@ -41,40 +40,76 @@ const CourseDashboardHeader = ({
   handleFilterChange,
   view,
   onViewChange,
-}: DashboardHeaderProps) => (
-  <Stack gap="md">
-    <Box>
-      <Title c="dark.7" variant="hero" order={2} fw={700}>
-        Courses
-      </Title>
-      <Text fw={500} c="dark.3" fz="md">
-        {/* Create a subtitle here */}
-        Browse and manage your courses
-      </Text>
-    </Box>
-    <Group align="center" justify="end" gap={rem(5)}>
-      <SearchComponent
-        data={coursesData}
-        onFilter={onSearchFilter}
-        identifiers={['courseName']}
-        placeholder="Search courses"
-      />
-      <CourseDashboardFilters
-        filters={filters}
-        activeFilters={activeFilters}
-        onAddFilter={handleAddFilter}
-        onRemoveFilter={handleRemoveFilter}
-        onFilterChange={handleFilterChange}
-      />
-      <ViewSelectorButton
-        view={view}
-        onGridClick={() => onViewChange('grid')}
-        onListClick={() => onViewChange('list')}
-      />
-    </Group>
-    <Divider />
-  </Stack>
-)
+}: DashboardHeaderProps) {
+  return (
+    <Stack gap="md">
+      <Box>
+        <Title c="dark.7" variant="hero" order={2} fw={700}>
+          Courses
+        </Title>
+        <Text fw={500} c="dark.3" fz="md">
+          {/* Create a subtitle here */}
+          Browse and manage your courses
+        </Text>
+      </Box>
+      <Group align="center" justify="end" gap={rem(5)}>
+        <SearchComponent
+          data={coursesData}
+          onFilter={onSearchFilter}
+          identifiers={['courseName']}
+          placeholder="Search courses"
+        />
+        <CourseDashboardFilters
+          filters={filters}
+          activeFilters={activeFilters}
+          onAddFilter={handleAddFilter}
+          onRemoveFilter={handleRemoveFilter}
+          onFilterChange={handleFilterChange}
+        />
+        <ViewSelectorButton
+          view={view}
+          onGridClick={() => onViewChange('grid')}
+          onListClick={() => onViewChange('list')}
+        />
+      </Group>
+      <Divider />
+    </Stack>
+  )
+}
+
+type CourseDashboardFiltersProps = {
+  filters: FilterType[]
+  activeFilters: FilterType[]
+  onAddFilter: (filterType: FilterType) => void
+  onRemoveFilter: (id: string) => void
+  onFilterChange: (id: string, value: string) => void
+}
+
+function CourseDashboardFilters({
+  filters,
+  activeFilters,
+  onAddFilter,
+  onRemoveFilter,
+  onFilterChange,
+}: CourseDashboardFiltersProps) {
+  const transformedFilters = useMemo(
+    () =>
+      activeFilters.map((filter) => ({
+        ...filter,
+        onChange: (value: string) => onFilterChange(filter.id, value),
+        onRemove: () => onRemoveFilter(filter.id),
+      })),
+    [activeFilters, onAddFilter, onFilterChange, onRemoveFilter],
+  )
+
+  return (
+    <MultiFilter
+      filters={filters}
+      activeFilters={transformedFilters}
+      onAddFilter={onAddFilter}
+    />
+  )
+}
 
 type FilterButtonProps = {
   showFilters: boolean
@@ -82,38 +117,43 @@ type FilterButtonProps = {
   onClick: () => void
 }
 
-const FilterButton = ({
+function FilterButton({
   showFilters,
   filterCount,
   onClick,
-}: FilterButtonProps) => (
-  <Tooltip label={showFilters ? 'Hide Filters' : 'Show Filters'}>
-    <Button
-      variant={'subtle'}
-      c={filterCount !== 0 || showFilters ? 'blue.6' : 'gray.6'}
-      onClick={onClick}
-      leftSection={<IconFilter2 size={20} />}
-    >
-      Filters
-    </Button>
-  </Tooltip>
-)
+}: FilterButtonProps) {
+  return (
+    <Tooltip label={showFilters ? 'Hide Filters' : 'Show Filters'}>
+      <Button
+        variant={'subtle'}
+        c={filterCount !== 0 || showFilters ? 'blue.6' : 'gray.6'}
+        onClick={onClick}
+        leftSection={<IconFilter2 size={20} />}
+      >
+        Filters
+      </Button>
+    </Tooltip>
+  )
+}
 
 type SelectorButtonProps = {
   active: boolean
   icon: ReactNode
   onClick: () => void
 }
-const SelectorButton = ({ active, icon, onClick }: SelectorButtonProps) => (
-  <Button
-    variant="default"
-    radius={'md'}
-    bg={active ? 'gray.3' : 'gray.0'}
-    onClick={onClick}
-  >
-    <Box color={active ? 'black' : 'dark.2'}>{icon}</Box>
-  </Button>
-)
+
+function SelectorButton({ active, icon, onClick }: SelectorButtonProps) {
+  return (
+    <Button
+      variant="default"
+      radius={'md'}
+      bg={active ? 'gray.3' : 'gray.0'}
+      onClick={onClick}
+    >
+      <Box color={active ? 'black' : 'dark.2'}>{icon}</Box>
+    </Button>
+  )
+}
 
 type ViewSelectorButtonProps = {
   view: 'grid' | 'list'
@@ -121,23 +161,25 @@ type ViewSelectorButtonProps = {
   onListClick: () => void
 }
 
-const ViewSelectorButton = ({
+function ViewSelectorButton({
   view,
   onGridClick,
   onListClick,
-}: ViewSelectorButtonProps) => (
-  <Button.Group>
-    <SelectorButton
-      active={view === 'grid'}
-      onClick={onGridClick}
-      icon={<IconLayoutGridFilled size={20} />}
-    />
-    <SelectorButton
-      active={view === 'list'}
-      onClick={onListClick}
-      icon={<IconList size={20} />}
-    />
-  </Button.Group>
-)
+}: ViewSelectorButtonProps) {
+  return (
+    <Button.Group>
+      <SelectorButton
+        active={view === 'grid'}
+        onClick={onGridClick}
+        icon={<IconLayoutGridFilled size={20} />}
+      />
+      <SelectorButton
+        active={view === 'list'}
+        onClick={onListClick}
+        icon={<IconList size={20} />}
+      />
+    </Button.Group>
+  )
+}
 
 export default CourseDashboardHeader
